Let both GET and POST autocomplete handle a json shape in the body.
GET can handle a body in HTTP, as is seen in e.g. Elasticsearch which we
use a lot. Some client libs have difficulty with this, so we should
leave the POST in place as well.

This makes both behave the same. Making the code a little easier by
removing some wrapper function callsLet both GET and POST autocomplete
handle a json shape in the body.

GET can handle a body in HTTP, as is seen in e.g. Elasticsearch which we
use a lot. Some client libs have difficulty with this, so we should
leave the POST in place as well.

This makes both behave the same. Making the code a little easier by
removing some wrapper function calls.