Move prometheus timer checking from rubber to query.
This way, only bragi initiates the query timer, and not other tooling
using the rubber.get_address such as poi2mimir and osm_reader.

It also allows us to label the query timer with "reverse", currently
the timer was unlabeled.

It furthermore allows a cleanup of some duplicated code.

I'm not certain, and it needs to be confirmed, but this appears to be
one cause of memory leaking in #412 as prometheus Metrics are maintained
in memory (are they always?)