Introduce a poi_config example file that includes most poi settings.
The POI configuration is adapted from
https://github.com/MorbZ/OsmPoisPbf/.

OsmPoisPbf has a filters.txt and a categories.csv. Both were combined
and ported to mimirsbrunn filter mechanism.

The human name is left uppercase so that implementers can translate this in
according to own locales or needs.

It uses wildcards as proposed in PR #415

This fixes #397.